+++ cpp/CMakeLists.txt
@@ -76,21 +76,22 @@ ExternalProject_Add(native-image
     DEPENDS mvn
     SOURCE_DIR ${PYPOWSYBL_JAVA_BIN_DIR}
     DOWNLOAD_COMMAND ""
-    PATCH_COMMAND $ENV{JAVA_HOME}/bin/native-image ${NATIVE_IMAGE_BUILD_OPTIONS} --class-path ${PYPOWSYBL_JAVA_SRC_DIR}/target/pypowsybl-java.jar${EXTRA_JARS} -march=compatibility --no-fallback --shared --gc=${NATIVE_IMAGE_GC} -H:Name=pypowsybl-java -H:CLibraryPath=${CMAKE_CURRENT_SOURCE_DIR}/src
+    PATCH_COMMAND $ENV{JAVA_HOME}/bin/native-image ${NATIVE_IMAGE_BUILD_OPTIONS} --class-path ${PYPOWSYBL_JAVA_SRC_DIR}/target/pypowsybl-java.jar${EXTRA_JARS} -march=compatibility --no-fallback --shared --gc=${NATIVE_IMAGE_GC} -H:Name=pypowsybl-java -H:CLibraryPath=${CMAKE_CURRENT_SOURCE_DIR}/powsybl-cpp
     CONFIGURE_COMMAND ""
     BUILD_COMMAND ""
     INSTALL_COMMAND ${CMAKE_COMMAND} -E copy ${PYPOWSYBL_JAVA_BIN_DIR}/${PYPOWSYBL_JAVA_OLD_LIB} ${PYPOWSYBL_JAVA_BIN_DIR}/${PYPOWSYBL_JAVA_LIB} ${NATIVE_IMAGE_INSTALL_EXTRA_COMMAND}
 )
 
-set(SOURCE_DIR "src")
+set(POWSYBL_CPP_SOURCE_DIR "powsybl-cpp")
+set(PYPOWSYBL_CPP_SOURCE_DIR "pypowsybl-cpp")
 
-include_directories(${SOURCE_DIR} ${PYPOWSYBL_JAVA_BIN_DIR})
-set(SOURCES "${SOURCE_DIR}/pypowsybl.cpp" "${SOURCE_DIR}/pylogging.cpp")
+include_directories(${POWSYBL_CPP_SOURCE_DIR} ${PYPOWSYBL_CPP_SOURCE_DIR} ${PYPOWSYBL_JAVA_BIN_DIR})
+set(SOURCES "${POWSYBL_CPP_SOURCE_DIR}/powsybl-cpp.cpp" "${PYPOWSYBL_CPP_SOURCE_DIR}/pylogging.cpp")
 
 link_directories(${PYPOWSYBL_JAVA_BIN_DIR})
 
 add_subdirectory(lib/pybind11)
-pybind11_add_module(_pypowsybl ${SOURCES} "${SOURCE_DIR}/bindings.cpp")
+pybind11_add_module(_pypowsybl ${SOURCES} "${PYPOWSYBL_CPP_SOURCE_DIR}/bindings.cpp")
 
 # extract powsybl math native from jar for current platform
 # powsybl-math-native.jar has been previously copied by maven build
