+++ cpp/powsybl-cpp/powsybl-cpp.cpp
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "pypowsybl.h"
+#include "powsybl-cpp.h"
 #include <iostream>
 
 namespace pypowsybl {
