+++ cpp/pypowsybl-cpp/pylogging.h
